fix: fix weird indentation in Subset
The Subset.ts files has problems when ESLint checks it. On some configurations the indentation seems
to be Ok, but in others, such as the one in the CI, is not. The file may be considered stable, and
it won't change through time. So we can safely change the configuration of ESLint to ingore the file
as to avoid future problems with it.